tmglow/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
blas=1.0=mkl
ca-certificates=2020.7.22=0
certifi=2020.6.20=py36_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py36_0
dataclasses=0.7=py36_0
dbus=1.13.16=hb2f20db_0
expat=2.2.9=he6710b0_2
fontconfig=2.13.0=h9420a91_0
freetype=2.10.2=h5ab3b9f_0
glib=2.65.0=h3eb4bd4_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
icu=58.2=he6710b0_3
intel-openmp=2020.2=254
jpeg=9b=h024ee3a_2
kiwisolver=1.2.0=py36hfd86e86_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=he19cac6_1
lz4-c=1.9.2=he6710b0_1
matplotlib=3.3.1=0
matplotlib-base=3.3.1=py36h817c723_0
mkl=2020.2=256
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.1.0=py36h23d657b_0
mkl_random=1.1.1=py36h0573a6f_0
ncurses=6.2=he6710b0_1
ninja=1.10.1=py36hfd86e86_0
numpy=1.19.1=py36hbc911f0_0
numpy-base=1.19.1=py36hfa32c7d_0
olefile=0.46=py36_0
openssl=1.1.1g=h7b6447c_0
pcre=8.44=he6710b0_0
pillow=7.2.0=py36hb39fc2d_0
pip=20.2.2=py36_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py36h05f1152_2
python=3.6.12=hcff3b4d_2
python-dateutil=2.8.1=py_0
python_abi=3.6=1_cp36m
pytorch=1.6.0=py3.6_cuda10.2.89_cudnn7.6.5_0
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
scipy=1.5.2=py36h0b6359f_0
setuptools=49.6.0=py36_0
sip=4.19.8=py36hf484d3e_0
six=1.15.0=py_0
sqlite=3.33.0=h62c20be_0
tk=8.6.10=hbc83047_0
tornado=6.0.4=py36h7b6447c_1
wheel=0.35.1=py_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
